PDF_QnA/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
distlib==0.3.7
exceptiongroup==1.2.0
faiss-cpu==1.7.4
filelock==3.12.4
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.3.3
google-api-core==2.14.0
google-auth==2.24.0
google-generativeai==0.2.2
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-status==1.59.3
idna==3.6
importlib-metadata==6.9.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.344
langchain-core==0.0.8
langsmith==0.0.68
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pipenv==2023.10.3
platformdirs==3.11.0
proto-plus==1.22.3
protobuf==4.25.1
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
streamlit==1.29.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
virtualenv==20.24.5
watchdog==3.0.0
yarl==1.9.3
zipp==3.17.0
